reestructurando a mvc datagrowth
