--- src/main/java/com/dgteam/dgblog/categoria/infra/CategoriaRepo.java
@@ -1,5 +1,0 @@
-package com.dgteam.dgblog.categoria.infra;
-
-public class CategoriaRepo {
-
-}
