--- src/main/java/com/dgteam/dgblog/categoria/model/ICategoriaRepo.java
@@ -1,5 +1,0 @@
-package com.dgteam.dgblog.categoria.model;
-
-public interface ICategoriaRepo {
-
-}
