+++ src/main/java/com/dgteam/dgblog/domain/service/IUsuarioService.java
@@ -1,0 +1,11 @@
+package com.dgteam.dgblog.domain.service;
+
+import java.util.List;
+
+
+import com.dgteam.dgblog.persistence.entity.Usuario;
+
+
+public interface IUsuarioService {
+	List<Usuario> getAll ();
+}
