--- src/main/java/com/dgteam/dgblog/users/domain/service/IUserService.java
@@ -1,11 +1,0 @@
-package com.dgteam.dgblog.users.domain.service;
-
-
-import com.dgteam.dgblog.users.domain.model.User;
-
-import java.util.Optional;
-
-public interface IUserService {
-    Optional<User> getUserById (Long id);
-
-}
